@@ -3,13 +3,18 @@
   "version": "3.0.1",
   "description": "A stylelint plugin that harnesses the power of postcss-bem-linter",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "lint": "eslint .",
-    "test": "jest"
+    "test": "NODE_OPTIONS=\"--experimental-vm-modules --no-warnings\" jest --runInBand"
   },
+  "exports": "./index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/simonsmith/stylelint-selector-bem-pattern.git"
+  },
+  "engines": {
+    "node": ">=18.12.0"
   },
   "files": [
     "index.js"
@@ -37,11 +42,11 @@
   "devDependencies": {
     "eslint": "^7.31.0",
     "jest": "^29.5.0",
-    "jest-preset-stylelint": "^6.1.0",
-    "stylelint": "^15.9.0"
+    "jest-preset-stylelint": "^7.0.0",
+    "stylelint": "^16.2.1"
   },
   "peerDependencies": {
-    "stylelint": ">=15.9.0 || 16"
+    "stylelint": "^16.2.1"
   },
   "jest": {
     "preset": "jest-preset-stylelint",
@@ -50,7 +55,7 @@
     ]
   },
   "volta": {
-    "node": "18.17.1",
+    "node": "20.10.0",
     "yarn": "1.22.19"
   }
 }
